modules/frontend/lambda-code/jwt-authorizer/requirements.txt
python-jose[cryptography]>=3.3.0